refactor workspace row markup into own component
